refactor: 타입별로 컴포넌트를 나눠서 순수하게 사용
